GUI v1.0 file format
