+++ src/View/Delivery/RemoveDeliveryView.form
